% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-tidywearable_update.R
\name{check_tidywearable_version}
\alias{check_tidywearable_version}
\title{check_tidywearable_version}
\usage{
check_tidywearable_version(
  packages = c("core", "all"),
  from = c("gitlab", "github", "gitee", "shen")
)
}
\arguments{
\item{packages}{core or all packages in tidywearable. "core" means all the core
packages in tidywearable and "all" means all the packages in tidywearable.}

\item{from}{"gitlab", "github", or "gitee", or "tidywearable.org"}
}
\description{
Check if there are packages in tidywearable can be updated.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
