% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-tidywearable_logo.R
\name{tidywearable_logo}
\alias{tidywearable_logo}
\title{Show the logo tidywearable.}
\usage{
tidywearable_logo(unicode = l10n_info()$`UTF-8`)
}
\arguments{
\item{unicode}{Whether to use Unicode symbols. Default is \code{TRUE}
on UTF-8 platforms.}
}
\value{
A ASCII log of tidywearable
}
\description{
The tidywearable logo, using ASCII or Unicode characters
}
\examples{
tidywearable_logo()
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
