% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-utils.R
\name{tidywearable_packages}
\alias{tidywearable_packages}
\title{List all packages in the tidywearable}
\usage{
tidywearable_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include tidywearable in the list?}
}
\description{
List all packages in the tidywearable
}
\examples{
tidywearable_packages()
}
