% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-tidywearable_update.R
\name{update_tidywearable}
\alias{update_tidywearable}
\title{update_tidywearable}
\usage{
update_tidywearable(
  packages = c("core", "all"),
  from = c("gitlab", "github", "gitee", "tidywearable.org"),
  fastgit = FALSE
)
}
\arguments{
\item{packages}{core or all packages in tidywearable. "core" means all the core
packages in tidywearable and "all" means all the packages in tidywearable.}

\item{from}{github, gitlab or gitee.}

\item{fastgit}{if install packages using fastgit. see
https://hub.fastgit.org/}
}
\description{
Update packages in tidywearable.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
